StockDatabaseManagement/requirements.txt
flask
mysql-connector-python
pandas
yfinance
pandas_datareader
python-dotenv
